You mast find the Annagrams
 What to do?
Words are called anagrams of each other if one of them can be obtained by rearranging the letters in the other. 
For example, the word "eat" can be obtained by rearranging the letters of the word "tea", so these words are anagrams of each other. 
Given a pair of words, check for each of them whether the words of this pair are anagrams of each other.