class fraction where you can do the standard operations like + - * /
I created a class of fractions which is reduced and the program also understands which fraction is more or less (this makes it possible to use map and set). 
You can also add, negate, multiply and divide two fractions(after the operation fraction would be reduced). It is possible to write fractions "Naturally" using operator overloading.
If numerator and denumerator is negative, program will understand this and fraction will be positive. Also, if the denumerator negative and the numerator is positive the fraction will be negative.
If numerator is 0. Fraction will be 0/1. If you don't write anything the fraction will be 0/1 too (standard fraction).