Database work
It is necessary to write a program in C ++ that implements working with a simple database (abbreviated as “DB”). The program will communicate with the user through standard input and output (stdin and stdout streams).

We will store in our database pairs of the form: date, event. We define the date as a string of the form Year-Month-Day, where Year, Month and Day are integers.

The event is defined as a string of arbitrary printed characters without separators inside (spaces, tabs, etc.). An event cannot be an empty string. Many different events can happen on the same date; the database should be able to save them all. The same events that occurred on the same day do not need to be saved: it is enough to save only one of them.

Our database must understand certain commands in order to interact with it:

A.
All commands, dates and events during input are separated by spaces. Commands are read from standard input. There can be exactly one command in one line, but you can enter several commands in several lines. The input may also receive empty lines - they should be ignored and continue processing new commands in subsequent lines.

Add Event (Add Date Event)
When adding an event, the database should remember it and then show it when searching (with the Find command) or printing (with the Print command). If the specified event for a given date already exists, re-adding it must be ignored. In case of correct input (see the section "Processing input errors"), the program should not display anything on the screen.

Delete Event (Del Date Event)
The team must remove the previously added event with the specified name on the specified date, if it exists. If the event is found and deleted, the program should display the string "Deleted successfully" (without quotes). If an event on the specified date is not found, the program should display the string “Event not found” (without quotes).

Delete multiple events (Del Date)
The command deletes all previously added events for the specified date. The program should always output a line of the form “Deleted N events”, where N is the number of all found and deleted events. N may be zero if there were no events on the specified date.

Event Search (Find Date)
Find and print previously added events on a specified date. The program should print only the events themselves, one per line. Events should be sorted in ascending order of string comparison between each other (type string).

Print all events (Print)
With this command you can show the full contents of our database. The program should print all the Date Event pairs, one per line. All pairs should be sorted by date, and events within the same date should be sorted in ascending order by which the strings are compared with each other (type string). Dates must be formatted in a special way: YYYY-MM-DD, where G, M, D are the digits of the numbers of the year, month and day, respectively. If a number has fewer digits, then it must be supplemented with zeros, for example, 0001-01-01 - the first of January of the first year. You do not need to display a date with a negative year value.